fix: move files around to prevent circular dependencies
