+++ packages/client/src/components/form/FormFieldGenerator.tsx
@@ -29,7 +29,6 @@ import {
   getFieldType,
   getQueryData,
   getVisibleOptions,
-  getListOfLocations,
   getFieldHelperText,
   getDependentFields,
   evalExpressionInFieldDefinition,
@@ -135,6 +134,7 @@ import { Heading2, Heading3 } from '@opencrvs/components/lib/Headings/Headings'
 import { SignatureUploader } from './SignatureField/SignatureUploader'
 import { ButtonField } from '@client/components/form/Button'
 import { RedirectField } from '@client/components/form/Redirect'
+import { getListOfLocations } from '@client/utils/validate'
 
 const SignatureField = styled(Stack)`
   margin-top: 8px;
