+++ packages/client/src/forms/utils.ts
@@ -510,16 +510,6 @@ export function isDefaultCountry(countryCode: string): boolean {
   return countryCode === window.config.COUNTRY.toUpperCase()
 }
 
-export function getListOfLocations(
-  resource: IOfflineData,
-  resourceType: Extract<
-    keyof IOfflineData,
-    'facilities' | 'locations' | 'offices'
-  >
-) {
-  return resource[resourceType]
-}
-
 interface IVars {
   [key: string]: any
 }
